added in-memory stub
