+++ build.gradle
@@ -86,6 +86,10 @@ subprojects {
             }
         }
 
+        test {
+            jvmArgs += '-Dvertx.disableDnsResolver=true'
+        }
+
     }
 
 }
