+++ infrastructure/spring-boot-webflux-rest-api-adapter/src/test/java/codes/fdk/blueprint/api/infrastructure/rest/webflux/ResetInMemoryRepoExtension.java
@@ -1,13 +1,13 @@
 package codes.fdk.blueprint.api.infrastructure.rest.webflux;
 
-import codes.fdk.blueprint.api.infrastructure.rest.webflux.RestApiWebfluxTestModule.InMemoryCategoryRepository;
+import codes.fdk.blueprint.api.domain.stub.InMemoryCategoryRepository;
 import org.junit.jupiter.api.extension.AfterEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 
 public class ResetInMemoryRepoExtension implements AfterEachCallback {
 
     @Override
-    public void afterEach(ExtensionContext context) throws Exception {
+    public void afterEach(ExtensionContext context) {
         InMemoryCategoryRepository.STORE.clear();
     }
