+++ infrastructure/spring-boot-webflux-rest-api-adapter/src/test/java/codes/fdk/blueprint/api/infrastructure/rest/webflux/RestApiWebfluxTestModule.java
@@ -1,18 +1,9 @@
 package codes.fdk.blueprint.api.infrastructure.rest.webflux;
 
-import codes.fdk.blueprint.api.domain.model.Category;
-import codes.fdk.blueprint.api.domain.model.CategoryId;
 import codes.fdk.blueprint.api.domain.service.CategoryService;
-import codes.fdk.blueprint.api.domain.spi.CategoryRepository;
+import codes.fdk.blueprint.api.domain.stub.InMemoryCategoryRepository;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-import java.util.UUID;
 
 @SpringBootApplication
 public class RestApiWebfluxTestModule {
@@ -20,56 +11,6 @@ public class RestApiWebfluxTestModule {
     @Bean
     public CategoryService categoryService() {
         return CategoryService.create(new InMemoryCategoryRepository());
-    }
-
-    static class InMemoryCategoryRepository implements CategoryRepository {
-
-        static final Map<CategoryId, Category> STORE = new HashMap<>();
-
-        @Override
-        public Mono<Category> save(Category category) {
-            return Mono.fromCallable(() -> {
-                Objects.requireNonNull(category);
-                final Category categoryWithId = assignIdWhenMissing(category);
-                return STORE.compute(categoryWithId.id(), (k, v) -> categoryWithId);
-            });
-        }
-
-        @Override
-        public Flux<Category> findAll() {
-            return Flux.fromIterable(STORE.values());
-        }
-
-        @Override
-        public Mono<Category> findById(CategoryId id) {
-            return Mono.defer(() -> {
-                Objects.requireNonNull(id);
-                return Mono.justOrEmpty(STORE.get(id));
-            });
-        }
-
-        @Override
-        public Flux<Category> findByParentId(CategoryId parentId) {
-            return Flux.defer(() -> {
-                Objects.requireNonNull(parentId);
-                return Flux.fromIterable(STORE.values())
-                           .filter(c -> parentId.equals(c.parentId()));
-            });
-        }
-
-        private static Category assignIdWhenMissing(Category category) {
-            if (category.id() == null) {
-                return new Category(
-                        CategoryId.of(UUID.randomUUID().toString()),
-                        category.name(),
-                        category.slug(),
-                        category.parentId(),
-                        category.isVisible()
-                );
-            }
-
-            return category;
-        }
     }
 
 }
